made content_archive modular, configured fronend to bundle via webpack to dist/, changed installers to load frontend from bundled dist/ instead of src/
