+++ ChatBot-Fin/Extension-ChatBot-Fin/check-dist.js
@@ -1,0 +1,28 @@
+// check-dist.js
+const fs = require('fs');
+const path = require('path');
+
+const distDir = path.resolve(__dirname, 'dist');
+const requiredFiles = ['manifest.json', 'bundle.js', 'textbox.css', 'assets/16x16_icon.png', 'assets/32x32_icon.png']; 
+
+let allGood = true;
+
+console.log(`\n🔍 Validating extension build in: ${distDir}\n`);
+
+for (const file of requiredFiles) {
+  const filePath = path.join(distDir, file);
+  if (!fs.existsSync(filePath)) {
+    console.error(`❌ Missing: ${file}`);
+    allGood = false;
+  } else {
+    console.log(`✅ Found: ${file}`);
+  }
+}
+
+if (allGood) {
+  console.log(`\n🎉 All required files are present in dist/. Ready to load into Chrome!\n`);
+  process.exit(0);
+} else {
+  console.error(`\n🚨 Some files are missing. Make sure Webpack and copy-webpack-plugin ran correctly.\n`);
+  process.exit(1);
+}
