+++ ChatBot-Fin/Extension-ChatBot-Fin/package.json
@@ -5,7 +5,9 @@
   "main": "src/content_archive.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack"
+    "build": "webpack",
+    "check": "node check-dist.js",
+    "build:full": "webpack && node check-dist.js"
   },
   "keywords": [],
   "author": "",
@@ -14,6 +16,7 @@
     "@babel/core": "^7.25.2",
     "@babel/preset-env": "^7.25.3",
     "babel-loader": "^9.1.3",
+    "copy-webpack-plugin": "^13.0.0",
     "webpack": "^5.94.0",
     "webpack-cli": "^5.1.4"
   }
