--- ChatBot-Fin/Extension-ChatBot-Fin/src/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>FinGPT Extension</title>
-    <link rel="stylesheet" href="./textbox.css">
-</head>
-<body>
-<script src="content_archive.js"></script>
-</body>
-</html>
