+++ ChatBot-Fin/Extension-ChatBot-Fin/src/modules/components/popup.js
@@ -1,0 +1,8 @@
+// popup.js
+function createPopup() {
+    const popup = document.createElement('div');
+    popup.id = "draggableElement";
+    return popup;
+}
+
+export { createPopup };
