+++ ChatBot-Fin/Extension-ChatBot-Fin/webpack.config.js
@@ -1,4 +1,5 @@
 const path = require('path');
+const CopyPlugin = require('copy-webpack-plugin');
 
 module.exports = {
     entry: './src/content_archive.js',
@@ -22,4 +23,14 @@ module.exports = {
     resolve: {
         extensions: ['.js'],
     },
+    plugins: [
+        new CopyPlugin({
+          patterns: [
+            { from: 'src/manifest.json', to: '.' },
+            { from: 'src/textbox.css', to: '.' },
+            { from: 'src/assets/', to: 'assets/' },
+            // Add other assets here if needed (icons, etc.)
+          ],
+        }),
+      ],
 };
